// apps/api/src/domain/exercise/exercise.entity.ts
export interface Exercise {
  id: string;
  slug: string;
  name: string;
  force: string | null;
  level: string;
  mechanic: string | null;
  equipment: string | null;
  primaryMuscles: string[];
  secondaryMuscles: string[];
  instructions: string[];
  category: string;
  images: string[];
  createdAt: Date;
}